// src/components/organisms/profile-tooltip/ProfileTooltip.types.ts
import type { ReactElement, HTMLAttributes } from 'react'
import { Address } from '../../../types'
import { ProfileExtraOptions } from '../full-width-profile/FullWidthProfile.types'
import { Boundary, FlipBehavior, TooltipPlacement } from '../../../hooks/useTooltipPosition'

export type ProfileTooltipExtraOptions = ProfileExtraOptions

export type ProfileCardProps = {
  addressOrName: Address | string
  connectedAddress?: Address
  darkMode?: boolean
  showSocials?: boolean
  showEmptySocials?: boolean
  showBio?: boolean
  showStatus?: boolean
  includeGrails?: boolean
  onProfileClick?: (addressOrName: Address | string) => void
  extraOptions?: ProfileTooltipExtraOptions
} & HTMLAttributes<HTMLDivElement>

export type ProfileTooltipProps = ProfileCardProps

export interface ProfileTooltipWrapperProps extends ProfileTooltipProps {
  children: ReactElement
  inline?: boolean
  verticalPlacement?: TooltipPlacement
  horizontalPlacement?: 'left' | 'right'
  verticalOffset?: number
  horizontalOffset?: number
  showArrow?: boolean
  showDelay?: number
  hideDelay?: number
  flipBehavior?: FlipBehavior
  boundary?: Boundary
  keepTooltipOnHover?: boolean
}